requirements.txt
# Snowflake-specific
snowflake-snowpark-python[pandas]==1.13.0
snowflake-ml-python==1.3.1

# to build Streamlit apps
streamlit==1.31.1
plotly==5.9.0
seaborn==0.12.0

# to connect to ChatGPT
openai==1.9.0
llama-index
llama-index-llms-openai
langchain
langchain-community
langchain-openai
snowflake-sqlalchemy

# Pandas Profiling
numba==0.58.1
ipywidgets
pandas-profiling
ydata-profiling

# TensorFlow and PyTorch
tensorflow
tf-keras
torch
torchdata
torch-utils

# miscellaneous
notebook
pypdf
faker
openpyxl
mlflow
transformers